create formsUtils and move resetLTS fnc into it
